requirements.txt
instabot==0.52.0
numpy==1.16.4
Pillow==6.1.0
python-dotenv==0.10.3
PyYAML==5.1.1
requests==2.22.0